Disable debug build for IBM Z and PPC arch

+++ image/rhel/download.sh
@@ -18,13 +18,17 @@ fi
 output_dir="/output"
 mkdir -p "${output_dir}/go/bin"
 if [[ "$DEBUG_BUILD" == "yes" ]]; then
-  dnf install -y golang
-  if [[ "$OSTYPE" != "linux-gnu"* ]]; then
-    GOBIN='' GOOS=linux GOARCH="${goarch}" GOPATH="${output_dir}/go" go install github.com/go-delve/delve/cmd/dlv@latest
-    mv "${output_dir}/go/bin/linux_${goarch}/dlv" "${output_dir}/go/bin/dlv"
-    rm -r "${output_dir}/go/bin/linux_${goarch}"
+  if [[ "$goarch" == "amd64" || "$goarch" == "arm64" ]]; then
+    dnf install -y golang
+    if [[ "$OSTYPE" != "linux-gnu"* ]]; then
+      GOBIN='' GOOS=linux GOARCH="${goarch}" GOPATH="${output_dir}/go" go install github.com/go-delve/delve/cmd/dlv@latest
+      mv "${output_dir}/go/bin/linux_${goarch}/dlv" "${output_dir}/go/bin/dlv"
+      rm -r "${output_dir}/go/bin/linux_${goarch}"
+    else
+      GOBIN="${output_dir}/go/bin" go install github.com/go-delve/delve/cmd/dlv@latest
+    fi
   else
-    GOBIN="${output_dir}/go/bin" go install github.com/go-delve/delve/cmd/dlv@latest
+    echo "WARNING: Architecture ${goarch} is not spported by delve. Rerun with DEBUG_BUILD=no"
   fi
 fi
